Proposal to include license information via #172.
Note that I am NOT the copyright owner or have any affiliation with the
piwik project.  This commit and the resulting PR were made informed by
aforementioned issue (#172) and the fact that the 'main' piwik git
repository includes this very repository and does not state
licence/copyright/trademark exceptions.  This commit is a pure proposal.