Make code compatible with Python <=2.7.8.
This fixes a regression introduced with PR #199 (commit a93a916e6efd).
The "context" parameter of the urllib2.HTTPSHandler constructor was only
introduced in Python 2.7.9, so we cannot pass a specific SSL context in older
Python versions. Luckily, we do not have to either because Python before 2.7.9
did not validate remote certificates anyway, so the behavior specified by the
--accept-invalid-ssl-certificate option is actually the default when running
on these versions of Python.