Add "--use-http-x-forwarded-for" option for NGINX
Updated the script to pick up an ip address from the `http_x_forwarded_for`
field instead of the `ip` field in NGINX JSON formatted log file provided that
the "--use-http-x-forwarded-for" option is used.  If the `http_x_forwarded_for`
field contains "-" or is missing, the script will pick up the ip address from
the `ip` field. This option is useful if the `ip` field does not contain
client's ip address (e.g. using NGINX as a reverse proxy).
Updated README file to include a field `http_x_forwarded_for` in NGINX
log file and information about the usage of "--use-http-x-forwarded-for" option.
Added two test cases to test the "--use-http-x-forwarded-for" option to ensure
that the script works properly.